+++ doc/changelog.md
@@ -4,6 +4,7 @@
 - (fix) crash on empty article list with article is selected (thanks to @rksvc)
 - (fix) invalid article title in RSS feeds with media containing titles (thanks to @bwwu-git for the report)
 - (fix) missing image enclosures in certain RSS feeds (thanks to @palinek for the report)
+- (fix) parsing namespaced legacy RSS feeds (thanks to @f100024)
 
 # v2.6 (2025-11-24)